+++ android/src/main/java/io/flutter/plugins/camera/CameraApiImpl.java
@@ -183,10 +183,13 @@ final class CameraApiImpl implements Messages.CameraApi {
     camera.capturePreviewFrame(result);
   }
   @Override
- public void capturePreviewFrameJpeg(@NonNull String outputPath, @NonNull Messages.Result<String> result) {
+  public void capturePreviewFrameJpeg(@NonNull String outputPath, @NonNull Messages.Result<String> result) {
     camera.capturePreviewFrameJpeg(outputPath,result);
   }
-
+  @Override
+  public void saveAsJpeg(@NonNull Map<String, Object> imageData, @NonNull String outputPath, @NonNull Long rotation, @NonNull Messages.Result<String> result){
+    camera.saveAsJpeg(imageData,outputPath,rotation.intValue(),result);
+  }
   @Override
   public void startListenFrames() {
     camera.startListenFrames(framesStreamChannel );