+++ android/src/main/java/io/flutter/plugins/camera/Messages.java
@@ -940,10 +940,12 @@ public class Messages {
      * resulting file.
      */
     void takePicture(@NonNull Result<String> result);
-
+    /** Capture a preview frame and return it as a map */
     void capturePreviewFrame(@NonNull Result<Map<String, Object>> result);
-
+    /** Capture a preview frame and return it as a jpeg */
     void capturePreviewFrameJpeg(@NonNull String outputPath, @NonNull Result<String> result);
+    /** Save a preview frame to the given path. */
+    void saveAsJpeg(@NonNull Map<String, Object> imageData, @NonNull String outputPath, @NonNull Long rotation, @NonNull Result<String> result);
     /** Start listening for preview frames */
     void startListenFrames();
     /** Stop listening for preview frames */
@@ -1242,6 +1244,37 @@ public class Messages {
                     };
 
                 api.capturePreviewFrameJpeg(outputPathArg, resultCallback);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger, "dev.flutter.pigeon.camera_android_frame.CameraApi.saveAsJpeg" + messageChannelSuffix, getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                ArrayList<Object> wrapped = new ArrayList<>();
+                ArrayList<Object> args = (ArrayList<Object>) message;
+                Map<String, Object> imageDataArg = (Map<String, Object>) args.get(0);
+                String outputPathArg = (String) args.get(1);
+                Long rotationArg = (Long) args.get(2);
+                Result<String> resultCallback =
+                    new Result<String>() {
+                      public void success(String result) {
+                        wrapped.add(0, result);
+                        reply.reply(wrapped);
+                      }
+
+                      public void error(Throwable error) {
+                        ArrayList<Object> wrappedError = wrapError(error);
+                        reply.reply(wrappedError);
+                      }
+                    };
+
+                api.saveAsJpeg(imageDataArg, outputPathArg, rotationArg, resultCallback);
               });
         } else {
           channel.setMessageHandler(null);