+++ lib/src/messages.g.dart
@@ -531,6 +531,7 @@ class CameraApi {
     }
   }
 
+  /// Capture a preview frame and return it as a map
   Future<Map<String, Object>> capturePreviewFrame() async {
     final String pigeonVar_channelName = 'dev.flutter.pigeon.camera_android_frame.CameraApi.capturePreviewFrame$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
@@ -558,6 +559,7 @@ class CameraApi {
     }
   }
 
+  /// Capture a preview frame and return it as a jpeg
   Future<String> capturePreviewFrameJpeg(String outputPath) async {
     final String pigeonVar_channelName = 'dev.flutter.pigeon.camera_android_frame.CameraApi.capturePreviewFrameJpeg$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
@@ -567,6 +569,34 @@ class CameraApi {
     );
     final List<Object?>? pigeonVar_replyList =
         await pigeonVar_channel.send(<Object?>[outputPath]) as List<Object?>?;
+    if (pigeonVar_replyList == null) {
+      throw _createConnectionError(pigeonVar_channelName);
+    } else if (pigeonVar_replyList.length > 1) {
+      throw PlatformException(
+        code: pigeonVar_replyList[0]! as String,
+        message: pigeonVar_replyList[1] as String?,
+        details: pigeonVar_replyList[2],
+      );
+    } else if (pigeonVar_replyList[0] == null) {
+      throw PlatformException(
+        code: 'null-error',
+        message: 'Host platform returned null value for non-null return value.',
+      );
+    } else {
+      return (pigeonVar_replyList[0] as String?)!;
+    }
+  }
+
+  /// Save a preview frame to the given path.
+  Future<String> saveAsJpeg(Map<String, Object> imageData, String outputPath, int rotation) async {
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.camera_android_frame.CameraApi.saveAsJpeg$pigeonVar_messageChannelSuffix';
+    final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: pigeonVar_binaryMessenger,
+    );
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_channel.send(<Object?>[imageData, outputPath, rotation]) as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {